split rules out into separate folders
